+++ CHANGELOG.md
@@ -166,9 +166,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
     - Length::from\_bytes() -> Length::try\_from\_bytes()
 - Make the following functions and methods unsafe
     - IdRef.as\_mut\_bytes()
+- `PartialEq::eq(Length::Indefinite, Length::Indefinite)` returns false, because they cannot be compared
 ### Removed
 - Delete the implementations for the following traits
     - Borrow<[u8]> for Id
     - Borrow<[u8]> for IdRef
     - Deref for IdRef
     - DerefMut for IdRef
+    - Eq for Length